Move network.py in seperate module
